Initial import
I've started stripping out the ex-template headers, and I've done a
little organisation, putting software pages into a software directory
and such.

Still lots to do.

I've also deleted all the old consortium pages, and some of the iridis
pages -- I kept the WinSCP Iridis page in case some of it was useful.
(Might re-cycle the images especially, but those aren't stored in here
yet.)

I'm thinking it doesn't really make sense to have a hojillion software
pages -- inevitably they'll become out of date and incomplete as we add
more software, and it's a lot of maintain when we make changes. I think
it might be worth looking through to see if any contain gnarly
work-arounds and deleting any that are just form "this is an
application that exists" pages.